Upgrade to Pulumi v2.12.0 and pulumi-terraform-bridge v2.11.0

+++ sdk/python/pulumi_wavefront/cloud_integration_cloud_trail.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from . import _utilities, _tables
 
 __all__ = ['CloudIntegrationCloudTrail']
@@ -25,7 +25,7 @@ class CloudIntegrationCloudTrail(pulumi.CustomResource):
                  region: Optional[pulumi.Input[str]] = None,
                  role_arn: Optional[pulumi.Input[str]] = None,
                  service: Optional[pulumi.Input[str]] = None,
-                 service_refresh_rate_in_minutes: Optional[pulumi.Input[float]] = None,
+                 service_refresh_rate_in_minutes: Optional[pulumi.Input[int]] = None,
                  __props__=None,
                  __name__=None,
                  __opts__=None):
@@ -59,7 +59,7 @@ class CloudIntegrationCloudTrail(pulumi.CustomResource):
         :param pulumi.Input[str] region: The AWS region of the S3 bucket where CloudTrail logs are stored
         :param pulumi.Input[str] role_arn: The external id corresponding to the Role ARN
         :param pulumi.Input[str] service: A value denoting which cloud service this service integrates with
-        :param pulumi.Input[float] service_refresh_rate_in_minutes: How often, in minutes, to refresh the service
+        :param pulumi.Input[int] service_refresh_rate_in_minutes: How often, in minutes, to refresh the service
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -119,7 +119,7 @@ class CloudIntegrationCloudTrail(pulumi.CustomResource):
             region: Optional[pulumi.Input[str]] = None,
             role_arn: Optional[pulumi.Input[str]] = None,
             service: Optional[pulumi.Input[str]] = None,
-            service_refresh_rate_in_minutes: Optional[pulumi.Input[float]] = None) -> 'CloudIntegrationCloudTrail':
+            service_refresh_rate_in_minutes: Optional[pulumi.Input[int]] = None) -> 'CloudIntegrationCloudTrail':
         """
         Get an existing CloudIntegrationCloudTrail resource's state with the given name, id, and optional extra
         properties used to qualify the lookup.
@@ -137,7 +137,7 @@ class CloudIntegrationCloudTrail(pulumi.CustomResource):
         :param pulumi.Input[str] region: The AWS region of the S3 bucket where CloudTrail logs are stored
         :param pulumi.Input[str] role_arn: The external id corresponding to the Role ARN
         :param pulumi.Input[str] service: A value denoting which cloud service this service integrates with
-        :param pulumi.Input[float] service_refresh_rate_in_minutes: How often, in minutes, to refresh the service
+        :param pulumi.Input[int] service_refresh_rate_in_minutes: How often, in minutes, to refresh the service
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
 
@@ -238,7 +238,7 @@ class CloudIntegrationCloudTrail(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="serviceRefreshRateInMinutes")
-    def service_refresh_rate_in_minutes(self) -> pulumi.Output[Optional[float]]:
+    def service_refresh_rate_in_minutes(self) -> pulumi.Output[Optional[int]]:
         """
         How often, in minutes, to refresh the service
         """
